Moved repo stuff out of router/admin and into repo/
