+++ repo/poll/poll_schedule.go
@@ -1,4 +1,4 @@
-package admin
+package poll
 
 import (
 	"context"
