+++ router/admin/ocelot_server.go
@@ -1,12 +1,13 @@
 package admin
 
 import (
+	"github.com/level11consulting/ocelot/repo"
 	"github.com/level11consulting/ocelot/secret"
 )
 
 type OcelotServer interface {
 	secret.SecretInterface
-	RepoInterface
+	repo.RepoInterface
 	StatusInterface
 	BuildInterface
 }
